[ YC-716 ] B2B customer accounts
- added setting for disabled addressbook management (this is different to B2B address book since the address book is allowed per customer but can only be entered from JAM by CC operative)
- aligned some of the summary screen settings with how we actually use those settings (some are at the customer shop level)

+++ web/support/src/main/java/org/yes/cart/web/support/service/impl/AddressBookFacadeImpl.java
@@ -221,7 +221,8 @@ public class AddressBookFacadeImpl implements AddressBookFacade {
 
     /** {@inheritDoc} */
     public void createOrUpdate(final Address address, final Shop customerShop) {
-        if (!customerShop.isB2BAddressBookActive()) {
+        if (!customerShop.isB2BAddressBookActive() &&
+                customerShop.isSfAddressBookEnabled(address.getCustomer().getCustomerType())) {
             if (address.getAddressId() == 0) {
                 // Need to add address to customer only just before the creation
                 address.getCustomer().getAddress().add(address);
@@ -235,7 +236,8 @@ public class AddressBookFacadeImpl implements AddressBookFacade {
     /** {@inheritDoc} */
     public void remove(Address address, final Shop customerShop) {
 
-        if (!customerShop.isB2BAddressBookActive()) {
+        if (!customerShop.isB2BAddressBookActive() &&
+                customerShop.isSfAddressBookEnabled(address.getCustomer().getCustomerType())) {
             final boolean isDefault = address.isDefaultAddress();
             addressService.delete(address);
 
@@ -253,7 +255,8 @@ public class AddressBookFacadeImpl implements AddressBookFacade {
 
     /** {@inheritDoc} */
     public Address useAsDefault(Address address, final Shop customerShop) {
-        if (!customerShop.isB2BAddressBookActive()) {
+        if (!customerShop.isB2BAddressBookActive() &&
+                customerShop.isSfAddressBookEnabled(address.getCustomer().getCustomerType())) {
             return addressService.updateSetDefault(address);
         }
         return address;
